refactor: 타입 경로 수정
